+++ src/base/static/js/views/place-detail-view.js
@@ -1,9 +1,11 @@
-/*globals Backbone _ jQuery Handlebars */
+  var Util = require('../utils.js');
 
-var Shareabouts = Shareabouts || {};
+  var SurveyView = require('mapseed-survey-view');
+  var SupportView = require('mapseed-support-view');
 
-(function(S, $, console){
-  S.PlaceDetailView = Backbone.View.extend({
+  var SubmissionCollection = require('../models/submission-collection.js');
+
+  module.exports = Backbone.View.extend({
     events: {
       'click .place-story-bar .btn-previous-story-nav': 'onClickStoryPrevious',
       'click .place-story-bar .btn-next-story-nav': 'onClickStoryNext'
@@ -13,30 +15,30 @@ var Shareabouts = Shareabouts || {};
 
       this.surveyType = this.options.surveyConfig.submission_type;
       this.supportType = this.options.supportConfig.submission_type;
-
+      
       this.model.on('change', this.onChange, this);
 
       // Make sure the submission collections are set
       this.model.submissionSets[this.surveyType] = this.model.submissionSets[this.surveyType] ||
-        new S.SubmissionCollection(null, {
+        new SubmissionCollection(null, {
           submissionType: this.surveyType,
           placeModel: this.model
         });
 
       this.model.submissionSets[this.supportType] = this.model.submissionSets[this.supportType] ||
-        new S.SubmissionCollection(null, {
+        new SubmissionCollection(null, {
           submissionType: this.supportType,
           placeModel: this.model
         });
 
-      this.surveyView = new S.SurveyView({
+      this.surveyView = new SurveyView({
         collection: this.model.submissionSets[this.surveyType],
         surveyConfig: this.options.surveyConfig,
         userToken: this.options.userToken,
         datasetId: self.options.datasetId
       });
 
-      this.supportView = new S.SupportView({
+      this.supportView = new SupportView({
         collection: this.model.submissionSets[this.supportType],
         supportConfig: this.options.supportConfig,
         userToken: this.options.userToken,
@@ -48,7 +50,7 @@ var Shareabouts = Shareabouts || {};
         // HACK! Each action should have its own view and bind its own events.
         var shareTo = this.getAttribute('data-shareto');
 
-        S.Util.log('USER', 'place', shareTo, self.model.getLoggingDetails());
+        Util.log('USER', 'place', shareTo, self.model.getLoggingDetails());
       });
     },
 
@@ -88,10 +90,9 @@ var Shareabouts = Shareabouts || {};
       this.delegateEvents();
 
       $("#content article").animate({ scrollTop: 0 }, "fast");
-      
+
       return this;
     },
-
     remove: function() {
       // Nothing yet
     },
@@ -100,4 +101,3 @@ var Shareabouts = Shareabouts || {};
       this.render();
     }
   });
-}(Shareabouts, jQuery, Shareabouts.Util.console));