+++ CHANGELOG.md
@@ -5,7 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - on our `develop` branch (unreleased)
+
+## [0.7.5] - 2017-03-05
+### Changed
+ - Replaced 'eval' statements with a static eval and AST parser
+ - Refactored pboakland flavor-specific views to extend from base views, instead of overriding the base views.
+
+## [0.7.4] - 2017-02-19
 ### Added
+ - webpack is now used for our bundling
+### Changed
+ - models.js has now been refactors into a models directory, separating all of our models.
 
 ## [0.7.3] - 2016-12-30
 ### Added
@@ -63,7 +73,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.6.0 - 2016-02-06
 
-[Unreleased]: https://github.com/smartercleanup/platform/compare/0.7.3...develop
+[Unreleased]: https://github.com/smartercleanup/platform/compare/0.7.5...develop
+[0.7.3]: https://github.com/smartercleanup/platform/compare/0.7.4...0.7.5
+[0.7.3]: https://github.com/smartercleanup/platform/compare/0.7.3...0.7.4
 [0.7.3]: https://github.com/smartercleanup/platform/compare/0.7.2...0.7.3
 [0.7.2]: https://github.com/smartercleanup/platform/compare/0.7.1...0.7.2
 [0.7.1]: https://github.com/smartercleanup/platform/compare/0.7.0...0.7.1